+++ components/Layout.tsx
@@ -16,7 +16,7 @@ const Layout = ({ children, title = "This is the default title" }: Props) => (
     </Head>
     <header>
       <nav>
-        <Link href="/">
+        <Link href="/home_page">
           <a>Home</a>
         </Link>{" "}
         |{" "}
