+++ pages/index.tsx
@@ -1,46 +1,25 @@
-import { useRef } from "react";
-import Layout from "../components/Layout";
+import Link from "next/link";
+import Head from "next/head";
+import React from "react";
 
-const IndexPage = () => {
-  const sendData = () =>
-    fetch("/api/data", {
-      method: "POST",
-      body: JSON.stringify({
-        email: refEmail?.current?.value,
-        password: refPassword?.current?.value,
-      }),
-    });
-
-  const refEmail = useRef<HTMLInputElement>(null);
-  const refPassword = useRef<HTMLInputElement>(null);
-
-  return (
-    <Layout title="Parking Lots">
-      <h1>Available spots</h1>
-      <ul>
-        <li>84.</li>
-        <li>85.</li>
-        <li>86.</li>
-        <li>87.</li>
-      </ul>
-
-      <input
-        type="email"
-        name="email"
-        ref={refEmail}
-        placeholder="test@example.com"
-      />
-      <br />
-      <input
-        type="password"
-        name="password"
-        ref={refPassword}
-        placeholder="password"
-      />
-      <br />
-      <button onClick={sendData}>Register</button>
-    </Layout>
-  );
-};
-
-export default IndexPage;
+function Login(){
+  return(
+<div>
+  <Head>
+    <title>Login page</title>
+  </Head>
+  <body>
+  <input
+    placeholder="email@example.com"
+  />
+  <br/>
+  <input
+    placeholder="password"
+  />
+  <br/>
+  <button>Login</button>
+  <br/>
+  Take me to <Link href="/register_page"> register </Link> new lemur
+  </body>
+</div>);}
+export default Login;
