+++ pages/style.css
@@ -1,0 +1,8 @@
+/* Register2ByJinV2 */
+
+div {
+  position: relative;
+  width: 1440px;
+  height: 1024px;
+  background: #edc7b7;
+}
